+++ test/helpers/index.js
@@ -3,11 +3,13 @@ import customFunctions from './customFunctions';
 import customImporter from './customImporter';
 import getCodeFromBundle from './getCodeFromBundle';
 import getCodeFromSass from './getCodeFromSass';
+import getCompiler from './getCompiler';
 import getErrors from './getErrors';
 import getImplementationByName from './getImplementationByName';
 import getTestId from './getTestId';
 import getWarnings from './getWarnings';
 import normalizeImplementationError from './normalizeImplementationError';
+import readAsset from './readAsset';
 
 export {
   compile,
@@ -15,9 +17,11 @@ export {
   customImporter,
   getCodeFromBundle,
   getCodeFromSass,
+  getCompiler,
   getErrors,
   getImplementationByName,
   getTestId,
   getWarnings,
   normalizeImplementationError,
+  readAsset,
 };
