+++ test/implementation-option.test.js
@@ -5,6 +5,7 @@ import Fiber from 'fibers';
 import {
   compile,
   getCodeFromBundle,
+  getCompiler,
   getErrors,
   getImplementationByName,
   getTestId,
@@ -32,8 +33,9 @@ describe('implementation option', () => {
       const options = {
         implementation: getImplementationByName(implementationName),
       };
-      const stats = await compile(testId, { loader: { options } });
-      const { css, sourceMap } = getCodeFromBundle(stats);
+      const compiler = getCompiler(testId, { loader: { options } });
+      const stats = await compile(compiler);
+      const { css, sourceMap } = getCodeFromBundle(stats, compiler);
 
       expect(css).toBeDefined();
       expect(sourceMap).toBeUndefined();
@@ -57,8 +59,9 @@ describe('implementation option', () => {
 
     const testId = getTestId('language', 'scss');
     const options = {};
-    const stats = await compile(testId, { loader: { options } });
-    const { css, sourceMap } = getCodeFromBundle(stats);
+    const compiler = getCompiler(testId, { loader: { options } });
+    const stats = await compile(compiler);
+    const { css, sourceMap } = getCodeFromBundle(stats, compiler);
 
     expect(css).toBeDefined();
     expect(sourceMap).toBeUndefined();
@@ -78,10 +81,12 @@ describe('implementation option', () => {
       }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -95,10 +100,12 @@ describe('implementation option', () => {
       }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -112,10 +119,12 @@ describe('implementation option', () => {
       }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -127,10 +136,12 @@ describe('implementation option', () => {
       implementation: Object.assign({}, dartSass, { info: 'asdfj' }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -142,10 +153,12 @@ describe('implementation option', () => {
       implementation: Object.assign({}, nodeSass, { info: 'node-sass\t1' }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -158,10 +171,12 @@ describe('implementation option', () => {
       implementation: Object.assign({}, dartSass, { info: undefined }),
     };
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(normalizeImplementationError(error)).toMatchSnapshot();
     }
@@ -189,10 +204,12 @@ describe('implementation option', () => {
     const testId = getTestId('language', 'scss');
     const options = {};
 
-    try {
-      const stats = await compile(testId, { loader: { options } });
+    const compiler = getCompiler(testId, { loader: { options } });
 
-      getCodeFromBundle(stats);
+    try {
+      const stats = await compile(compiler);
+
+      getCodeFromBundle(stats, compiler);
     } catch (error) {
       expect(error).toMatchSnapshot();
     }
